# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TalentLock API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employer-documents/upload-url:
    post:
      operationId: postEmployerDocumentsUploadUrl
      tags: [employer-documents]
      summary: Request an employer business document upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployerDocumentUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerDocumentUploadUrl"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
        "403":
          description: Employer profile required
  /employer-documents/confirm:
    post:
      operationId: postEmployerDocumentsConfirm
      tags: [employer-documents]
      summary: Confirm an employer business document upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployerDocumentConfirmBody"
      responses:
        "201":
          description: Document queued for review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerDocumentConfirmResponse"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "403":
          description: Employer profile required
  /employer-documents/me:
    get:
      operationId: getEmployerDocumentsMe
      tags: [employer-documents]
      summary: Get the current employer's safe document statuses
      responses:
        "200":
          description: Employer document statuses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerDocumentsMeResponse"
        "401":
          description: Unauthorized
        "403":
          description: Employer profile required
  /employer-documents/me/{documentType}/view-url:
    get:
      operationId: getEmployerDocumentsMeDocumentTypeViewUrl
      tags: [employer-documents]
      summary: Get a short-lived URL to view an own employer document
      parameters:
        - name: documentType
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/EmployerDocumentType"
      responses:
        "200":
          description: Signed view URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerDocumentViewUrl"
        "401":
          description: Unauthorized
        "403":
          description: Employer profile required
        "404":
          description: Document not found
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── User ──────────────────────────────────────────────────────────────────
...
    EmployerDocumentType:
      type: string
      enum:
        - company_registration
        - tax_vat_certificate
        - business_licence
        - representative_id
        - proof_of_business_address
    EmployerDocumentUploadUrlBody:
      type: object
      properties:
        documentType:
          $ref: "#/components/schemas/EmployerDocumentType"
        filename:
          type: string
        mimeType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
      required: [documentType, filename, mimeType]
    EmployerDocumentUploadUrl:
      type: object
      properties:
        uploadUrl:
          type: string
        fileUrl:
          type: string
      required: [uploadUrl, fileUrl]
    EmployerDocumentConfirmBody:
      type: object
      properties:
        documentType:
          $ref: "#/components/schemas/EmployerDocumentType"
        fileUrl:
          type: string
      required: [documentType, fileUrl]
    EmployerDocumentConfirmResponse:
      type: object
      properties:
        documentId:
          type: integer
        status:
          type: string
          enum: [pending]
      required: [documentId, status]
    EmployerDocumentStatus:
      type: object
      properties:
        documentType:
          $ref: "#/components/schemas/EmployerDocumentType"
        status:
          type: string
          enum: [pending, verified, rejected, needs_review]
        employerNotes:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
      required: [documentType, status, employerNotes, updatedAt]
    EmployerDocumentsMeResponse:
      type: object
      properties:
        verificationLevel:
          type: string
          enum: [unverified, partially_verified, fully_verified]
        isVerified:
          type: boolean
        documents:
          type: array
          items:
            $ref: "#/components/schemas/EmployerDocumentStatus"
      required: [verificationLevel, isVerified, documents]
    EmployerDocumentViewUrl:
      type: object
      properties:
        signedUrl:
          type: string
        expiresInSeconds:
          type: integer
      required: [signedUrl, expiresInSeconds]
...