# lib/api-zod/src/generated/types/employerDocumentType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TalentLock API specification
 * OpenAPI spec version: 0.1.0
 */

export type EmployerDocumentType =
  (typeof EmployerDocumentType)[keyof typeof EmployerDocumentType];

export const EmployerDocumentType = {
  company_registration: "company_registration",
  tax_vat_certificate: "tax_vat_certificate",
  business_licence: "business_licence",
  representative_id: "representative_id",
  proof_of_business_address: "proof_of_business_address",
} as const;