# lib/api-zod/src/generated/types/employerDocumentsMeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TalentLock API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmployerDocumentsMeResponseVerificationLevel } from "./employerDocumentsMeResponseVerificationLevel";
import type { EmployerDocumentStatus } from "./employerDocumentStatus";

export interface EmployerDocumentsMeResponse {
  verificationLevel: EmployerDocumentsMeResponseVerificationLevel;
  isVerified: boolean;
  documents: EmployerDocumentStatus[];
}